+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Ancestors & descendants".
 - Tests for "Paving with bricks".
 - Tests for "Playfair cipher".
+- Tests for "Barcode scanner".
 
 ## [1.14.0] - 2022-10-31
 ### Added